Add files via upload
This is the basic outline of the homepage. As the bottom part of the webpage will be used repeatedly, I've stored the file in the snippets folder to use whenever needed. I also separated the code of the options so that we can use them whenever we need them on another webpage.